==> susquehanna/notebooks/10. Performance metrics table.ipynb <==
import sys
import pandas as pd

sys.path.append("..")
import os
import rbf_functions

rbfs = [
    rbf_functions.original_rbf,
    rbf_functions.squared_exponential_rbf,
    rbf_functions.inverse_quadratic_rbf,
    rbf_functions.inverse_multiquadric_rbf,
    rbf_functions.exponential_rbf,
    rbf_functions.matern32_rbf,
    rbf_functions.matern52_rbf,
]

# Load data

hv_global = pd.read_csv("./calculated_metrics/hv_refset_global.csv").iloc[0, -1]
averages = {}
for entry in rbfs:
    rbf = entry.__name__

    # normalize hv against reference set so reference set is 1
    df_hv = pd.read_csv(f"calculated_metrics/hv_{rbf}_global.csv")
    df_hv.hypervolume /= hv_global

    df_ei = pd.read_csv(f"calculated_metrics/ei_{rbf}_global.csv")
    df_gd = pd.read_csv(f"calculated_metrics/gd_{rbf}_global.csv")

    avg_hv = df_hv.groupby("seed").tail(1).loc[:, "hypervolume"].mean()
    avg_ei = df_ei.groupby("ei").tail(1).loc[:, "ei"].mean()
    avg_gd = df_gd.groupby("gd").tail(1).loc[:, "gd"].mean()
    averages[rbf] = dict(
        hypervolume=avg_hv, epsilon_indicator=avg_ei, generational_distance=avg_gd
    )

scores = pd.DataFrame.from_dict(averages).T
scores

import os

paretosets = {}
size = {}
for entry in rbfs:
    name = entry.__name__
    # load variables
    varlist = []
    variables = []

    output_dir = os.path.abspath("./refsets/")
    for filename in os.listdir(output_dir):
        if filename == f"{name}_refset.csv":
            varlist.append(filename[:-4])
            df_vars = pd.read_csv(os.path.join(output_dir, filename))
            paretosets[name] = df_vars
            size[name] = len(df_vars)
    variables = df_vars.values
#     print(f"Loaded: {', '.join(varlist)}")

size

scores["nr. of solutions"] = pd.Series(size)

scores

# set contributions

df_global = pd.read_csv(f"./refsets/global_refset.csv")

l_global = df_global.values.tolist()
contribution = {}
templ = []
for rbf in paretosets:
    x = 0
    for solution in paretosets[rbf].values.tolist():
        if solution in l_global:
            templ.append(solution)
            x += 1
    contribution[rbf] = round(x / len(df_global), 3)
    rbfd[rbf] = pd.DataFrame(templ, columns=data.columns)
    templ = []
contribution

scores["set contribution"] = pd.Series(contribution)

scores

scores.to_latex()

scores.cols

scores[
    [
        "nr. of solutions",
        "generational_distance",
        "epsilon_indicator",
        "hypervolume",
        "set contribution",
    ]
].to_latex()
