==> susquehanna/notebooks/11. Decision variable analysis.ipynb <==
import sys
import numpy as np
import pandas as pd
import os

sys.path.append("..")
import rbf_functions

rbfs = [
    rbf_functions.original_rbf,
    rbf_functions.squared_exponential_rbf,
    rbf_functions.inverse_quadratic_rbf,
    rbf_functions.inverse_multiquadric_rbf,
    rbf_functions.exponential_rbf,
    rbf_functions.matern32_rbf,
    rbf_functions.matern52_rbf,
]

output_dir = os.path.abspath("../notebooks/refsets/")
name = "original_rbf"
filename = f"{name}_refset_variables.csv"
filename2 = f"{name}_refset.csv"
df_dvars = pd.read_csv(f"{output_dir}\{filename}", header=None)
df_refset = pd.read_csv(f"{output_dir}\{filename2}")

columnname = [
    "c1",
    "r1",
    "c2",
    "r2",
    "c3",
    "r3",
    "c4",
    "r4",
    "c5",
    "r5",
    "c6",
    "r6",
    "c7",
    "r7",
    "c8",
    "r8",
    "w1",
    "w2",
    "w3",
    "w4",
    "w5",
    "w6",
    "w7",
    "w8",
    "w9",
    "w10",
    "w11",
    "w12",
    "w13",
    "w14",
    "w15",
    "w16",
]
df_dvars.columns = columnname
df_dvars.describe()

df_dvars[["r1", "r2", "r3", "r4", "r5", "r6", "r7", "r8"]].describe().round(2)

# Original
output_dir = os.path.abspath("../notebooks/refsets/")
name = "original_rbf"
filename = f"{name}_refset_variables.csv"
df_dvars = pd.read_csv(f"{output_dir}\{filename}", header=None)
df_dvars.columns = columnname

df = pd.DataFrame(columns=["center", "radius", "weights"])
df["center"] = (
    df_dvars["c1"]
    + df_dvars["c2"]
    + df_dvars["c3"]
    + df_dvars["c4"]
    + df_dvars["c5"]
    + df_dvars["c6"]
    + df_dvars["c7"]
    + df_dvars["c8"]
) / 8
df["radius"] = (
    df_dvars["r1"]
    + df_dvars["r2"]
    + df_dvars["r3"]
    + df_dvars["r4"]
    + df_dvars["r5"]
    + df_dvars["r6"]
    + df_dvars["r7"]
    + df_dvars["r8"]
) / 8
df["weight"] = (
    df_dvars["w1"]
    + df_dvars["w2"]
    + df_dvars["w3"]
    + df_dvars["w4"]
    + df_dvars["w5"]
    + df_dvars["w6"]
    + df_dvars["w7"]
    + df_dvars["w8"]
    + df_dvars["w9"]
    + df_dvars["w10"]
    + df_dvars["w11"]
    + df_dvars["w12"]
    + df_dvars["w13"]
    + df_dvars["w14"]
    + df_dvars["w15"]
    + df_dvars["w16"]
) / 16
# df.loc[:,"center"].map('{:.2f}'.format)
df.describe().round(2)

# IMQ
output_dir = os.path.abspath("../notebooks/refsets/")
name = "inverse_multiquadric_rbf"
filename = f"{name}_refset_variables.csv"
df_dvars = pd.read_csv(f"{output_dir}\{filename}", header=None)
df_dvars.columns = columnname

df = pd.DataFrame(columns=["center", "radius", "weights"])
df["center"] = (
    df_dvars["c1"]
    + df_dvars["c2"]
    + df_dvars["c3"]
    + df_dvars["c4"]
    + df_dvars["c5"]
    + df_dvars["c6"]
    + df_dvars["c7"]
    + df_dvars["c8"]
) / 8
df["radius"] = (
    df_dvars["r1"]
    + df_dvars["r2"]
    + df_dvars["r3"]
    + df_dvars["r4"]
    + df_dvars["r5"]
    + df_dvars["r6"]
    + df_dvars["r7"]
    + df_dvars["r8"]
) / 8
df["weight"] = (
    df_dvars["w1"]
    + df_dvars["w2"]
    + df_dvars["w3"]
    + df_dvars["w4"]
    + df_dvars["w5"]
    + df_dvars["w6"]
    + df_dvars["w7"]
    + df_dvars["w8"]
    + df_dvars["w9"]
    + df_dvars["w10"]
    + df_dvars["w11"]
    + df_dvars["w12"]
    + df_dvars["w13"]
    + df_dvars["w14"]
    + df_dvars["w15"]
    + df_dvars["w16"]
) / 16
# df.loc[:,"center"].map('{:.2f}'.format)
df.describe().round(2)

# Exponential
output_dir = os.path.abspath("../notebooks/refsets/")
name = "exponential_rbf"
filename = f"{name}_refset_variables.csv"
df_dvars = pd.read_csv(f"{output_dir}\{filename}", header=None)
df_dvars.columns = columnname

df = pd.DataFrame(columns=["center", "radius", "weights"])
df["center"] = (
    df_dvars["c1"]
    + df_dvars["c2"]
    + df_dvars["c3"]
    + df_dvars["c4"]
    + df_dvars["c5"]
    + df_dvars["c6"]
    + df_dvars["c7"]
    + df_dvars["c8"]
) / 8
df["radius"] = (
    df_dvars["r1"]
    + df_dvars["r2"]
    + df_dvars["r3"]
    + df_dvars["r4"]
    + df_dvars["r5"]
    + df_dvars["r6"]
    + df_dvars["r7"]
    + df_dvars["r8"]
) / 8
df["weight"] = (
    df_dvars["w1"]
    + df_dvars["w2"]
    + df_dvars["w3"]
    + df_dvars["w4"]
    + df_dvars["w5"]
    + df_dvars["w6"]
    + df_dvars["w7"]
    + df_dvars["w8"]
    + df_dvars["w9"]
    + df_dvars["w10"]
    + df_dvars["w11"]
    + df_dvars["w12"]
    + df_dvars["w13"]
    + df_dvars["w14"]
    + df_dvars["w15"]
    + df_dvars["w16"]
) / 16
df.describe().round(2)
